requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
Pillow==12.2.0
gunicorn==25.3.0
